Orquestrador/requirements.txt
fastapi

uvicorn

httpx
pywebview

pyyaml
